feat(embedded/sql): sql transaction creation with options
