+++ embedded/sql/grouped_row_reader_test.go
@@ -34,7 +34,7 @@ func TestGroupedRowReader(t *testing.T) {
 	_, err = newGroupedRowReader(nil, nil, nil)
 	require.Equal(t, ErrIllegalArguments, err)
 
-	tx, err := engine.NewTx(context.Background())
+	tx, err := engine.NewTx(context.Background(), DefaultTxOptions())
 	require.NoError(t, err)
 
 	db, err := tx.catalog.newDatabase(1, "db1")
