+++ embedded/sql/joint_row_reader_test.go
@@ -35,7 +35,7 @@ func TestJointRowReader(t *testing.T) {
 	_, err = newJointRowReader(nil, nil)
 	require.Equal(t, ErrIllegalArguments, err)
 
-	tx, err := engine.NewTx(context.Background())
+	tx, err := engine.NewTx(context.Background(), DefaultTxOptions())
 	require.NoError(t, err)
 
 	db, err := tx.catalog.newDatabase(1, "db1")
